+++ jena-arq/src/test/java/org/apache/jena/riot/writer/TestJsonLDWriter.java
@@ -38,7 +38,6 @@ import org.apache.jena.riot.JsonLDWriteContext;
 import org.apache.jena.riot.RDFDataMgr;
 import org.apache.jena.riot.RDFFormat;
 import org.apache.jena.riot.WriterDatasetRIOT;
-import org.apache.jena.riot.out.JsonLDWriter;
 import org.apache.jena.riot.system.PrefixMap;
 import org.apache.jena.riot.system.RiotLib;
 import org.apache.jena.sparql.core.DatasetGraph;
